Partially revert "Delete unnecessary options"
This reverts commit 0aa7307, but with the
addition that the given options are hidden
help page. The reason is, that --buf-cap
is used in an unit test and the other options
may be used as well in the future.
The reader buffer size is kept at 64KiB
(not 68KiB as before), since 64KiB is the
current default of seq_io.